compare dfs functions
