+++ mssql_dataframe/__equality__.py
@@ -1,0 +1,12 @@
+# TODO: replace .equals all tests
+def compare_dfs(df1, df2):
+    
+    if df1.equals(df2):
+        return True
+    
+    assert df1.columns.equals(df2.columns)
+    
+    assert df1.index.equals(df2.index)
+    
+    for col in df1.columns:
+        assert df1[col].equals(df2[col])
