+++ tests/test_core/test_create.py
@@ -9,6 +9,7 @@ import pyodbc
 
 from mssql_dataframe.connect import connect
 from mssql_dataframe.core import conversion, create
+from mssql_dataframe.__equality__ import compare_dfs
 
 pd.options.mode.chained_assignment = "raise"
 
@@ -20,20 +21,6 @@ class package:
         self.create_meta = create.create(
             self.connection, include_metadata_timestamps=True
         )
-
-
-# TODO: replace .equals in other tests
-def compare_dfs(df1, df2):
-    
-    if df1.equals(df2):
-        return True
-    
-    assert df1.columns.equals(df2.columns)
-    
-    assert df1.index.equals(df2.index)
-    
-    for col in df1.columns:
-        assert df1[col].equals(df2[col])
 
 
 @pytest.fixture(scope="module")
